// src/payload-types.ts
export interface User {
  id: number;
  name?: string | null;
  address?: string | null;
  city?: string | null;
  cp?: string | null;
  province?:
    | (
        | 'Alava'
        | 'Albacete'
        | 'Alicante'
        | 'Almeria'
        | 'Asturias'
        | 'Avila'
        | 'Badajoz'
        | 'Barcelona'
        | 'Burgos'
        | 'Caceres'
        | 'Cadiz'
        | 'Cantabria'
        | 'Castellon'
        | 'Ciudad_Real'
        | 'Cordoba'
        | 'Cuenca'
        | 'Gerona'
        | 'Granada'
        | 'Guadalajara'
        | 'Guipuzcoa'
        | 'Huelva'
        | 'Huesca'
        | 'Baleares'
        | 'Jaen'
        | 'La_Coruna'
        | 'La_Rioja'
        | 'Las_Palmas'
        | 'Leon'
        | 'Lerida'
        | 'Lugo'
        | 'Madrid'
        | 'Malaga'
        | 'Murcia'
        | 'Navarra'
        | 'Orense'
        | 'Palencia'
        | 'Pontevedra'
        | 'Salamanca'
        | 'Tenerife'
        | 'Segovia'
        | 'Sevilla'
        | 'Soria'
        | 'Tarragona'
        | 'Teruel'
        | 'Toledo'
        | 'Valencia'
        | 'Valladolid'
        | 'Vizcaya'
        | 'Zamora'
        | 'Zaragoza'
      )
    | null;
  mailing_name?: string | null;
  mailing_address?: string | null;
  mailing_city?: string | null;
  mailing_cp?: string | null;
  mailing_province?:
    | (
        | 'Alava'
        | 'Albacete'
        | 'Alicante'
        | 'Almeria'
        | 'Asturias'
        | 'Avila'
        | 'Badajoz'
        | 'Barcelona'
        | 'Burgos'
        | 'Caceres'
        | 'Cadiz'
        | 'Cantabria'
        | 'Castellon'
        | 'Ciudad_Real'
        | 'Cordoba'
        | 'Cuenca'
        | 'Gerona'
        | 'Granada'
        | 'Guadalajara'
        | 'Guipuzcoa'
        | 'Huelva'
        | 'Huesca'
        | 'Baleares'
        | 'Jaen'
        | 'La_Coruna'
        | 'La_Rioja'
        | 'Las_Palmas'
        | 'Leon'
        | 'Lerida'
        | 'Lugo'
        | 'Madrid'
        | 'Malaga'
        | 'Murcia'
        | 'Navarra'
        | 'Orense'
        | 'Palencia'
        | 'Pontevedra'
        | 'Salamanca'
        | 'Tenerife'
        | 'Segovia'
        | 'Sevilla'
        | 'Soria'
        | 'Tarragona'
        | 'Teruel'
        | 'Toledo'
        | 'Valencia'
        | 'Valladolid'
        | 'Vizcaya'
        | 'Zamora'
        | 'Zaragoza'
      )
    | null;
  cif_dni_nie?: string | null;
  role: 'admin' | 'user' | 'pro';
  activated: boolean;
  /**
   * Marcar para enviar email de bienvenida al activar cuenta PRO
   */
  sendProEmail?: boolean | null;
  certificate?: (number | null) | Certificate;
  welcomeEmailSent?: boolean | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "certificate".
 */
export interface Certificate {
  id: number;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  title: string;
  description?: string | null;
  image?: (number | null) | Media;
  slug?: string | null;
  slugLock?: boolean | null;
  parent?: (number | null) | Category;
  subcategories?: (number | Category)[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "brands".
 */
export interface Brand {
  id: number;
  title: string;
  description?: string | null;
  image?: (number | null) | Media;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tags".
 */
export interface Tag {
  id: number;
  title: string;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: number;
  title: string;
  description?: string | null;
  price: number;
  price_wholesale: number;
  stock: number;
  image?: (number | null) | Media;
  images?: ImageBlock[] | null;
  slug?: string | null;
  slugLock?: boolean | null;
  brand: number | Brand;
  tags?: (number | Tag)[] | null;
  categories: (number | Category)[];
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ImageBlock".
 */
export interface ImageBlock {
  img?: (number | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'image';
}